Split Fal webhook refund helpers
